Fix: Add doxygen in the ft_string.  Changes to be committed: 	modified:   src/ft_string/ft_bzero.c 	modified:   src/ft_string/ft_memchr.c 	modified:   src/ft_string/ft_memcmp.c 	modified:   src/ft_string/ft_memcpy.c 	modified:   src/ft_string/ft_memmove.c 	modified:   src/ft_string/ft_memset.c 	modified:   src/ft_string/ft_strchr.c 	modified:   src/ft_string/ft_strdup.c 	modified:   src/ft_string/ft_strlcat.c 	modified:   src/ft_string/ft_strlcpy.c 	modified:   src/ft_string/ft_strlen.c 	modified:   src/ft_string/ft_strncmp.c 	modified:   src/ft_string/ft_strnstr.c 	modified:   src/ft_string/ft_strrchr.c

+++ src/ft_string/ft_memcmp.c
@@ -6,20 +6,34 @@
 /*   By: dande-je <dande-je@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/17 23:01:53 by dande-je          #+#    #+#             */
-/*   Updated: 2023/12/09 21:04:07 by dande-je         ###   ########.fr       */
+/*   Updated: 2024/01/22 01:03:41 by dande-je         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../include/libft.h"
+#include "ft_string.h"
 
+/**
+ * @brief Compares the first `n` bytes (each interpreted as unsigned char) of
+ * the memory areas `s1` and `s2`.
+ *
+ * @param s1 A pointer to the first memory area.
+ * @param s2 A pointer to the second memory area.
+ * @param n The number of bytes to be compared.
+ * @return An integer less than, equal to, or greater than zero if the first `n`
+ * bytes of `s1` is found, respectively, to be less than, to match, or be
+ * greater than the first `n` bytes of `s2`. For a nonzero return value, the
+ * sign is determined by the sign of the difference between the first pair of
+ * bytes (interpreted as unsigned char) that differ in `s1` and `s2`. If `n` is
+ * zero, the return value is zero.
+ */
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	long int	i;
+	size_t	i;
 
 	i = -1;
 	if (n == 0)
 		return (0);
-	while (++i < (long int)n)
+	while (++i < n)
 		if (*(unsigned char *)(s1 + i) != *(unsigned char *)(s2 + i))
 			return (*(unsigned char *)(s1 + i) - *(unsigned char *)(s2 + i));
 	return (0);
