+++ src/ft_string/ft_memcpy.c
@@ -6,20 +6,28 @@
 /*   By: dande-je <dande-je@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/17 00:13:40 by dande-je          #+#    #+#             */
-/*   Updated: 2023/12/09 21:04:09 by dande-je         ###   ########.fr       */
+/*   Updated: 2024/01/22 03:57:42 by dande-je         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../include/libft.h"
+#include "ft_string.h"
 
+/**
+ * @brief Copies `n` bytes from memory area `src` to memory area `dest`.
+ *
+ * @param dest A pointer to the destination memory area.
+ * @param src A pointer to the source memory area.
+ * @param n The number of bytes to be copied.
+ * @return A pointer to `dest`.
+ */
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	long int	i;
+	size_t	i;
 
 	i = -1;
 	if (!dest && !src)
 		return (NULL);
-	while (++i < (long int)n)
+	while (++i < n)
 		*(char *)(dest + i) = *(char *)(src + i);
 	return (dest);
 }
