+++ src/ft_string/ft_strlen.c
@@ -6,15 +6,22 @@
 /*   By: dande-je <dande-je@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/24 19:37:15 by dande-je          #+#    #+#             */
-/*   Updated: 2023/12/09 21:04:28 by dande-je         ###   ########.fr       */
+/*   Updated: 2024/01/22 04:46:08 by dande-je         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../include/libft.h"
+#include "ft_string.h"
 
+/**
+ * @brief Calculates the length of the string pointed to by `s`, excluding the
+ * terminating null byte ('\0').
+ *
+ * @param s The string to be scanned.
+ * @return Number of bytes in the string pointed to by `s`.
+ */
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
